requirements.txt
flask
numpy==1.15.4
keras==2.1.2
tensorflow==1.12.0
opencv-python
imutils
Pillow